+++ .firebaserc
@@ -1,5 +1,5 @@
 {
   "projects": {
-    "default": "cse112-team1"
+    "default": "chrome-extension-cse-112"
   }
 }
