+++ .github/workflows/Deploy.yml
@@ -1,8 +1,10 @@
 name: Build and Deploy
+
 on:
-  push:
+  pull_request:
     branches:
       - master
+      - develop
 
 jobs:
   build:
