+++ __mocks__/chromeMock.js
@@ -1,0 +1,10 @@
+let identity = {
+  getProfileUserInfo: jest.fn((callback) => {
+    callback({ email: "test@gmail.com" });
+  }),
+  getRedirectURL: jest.fn(),
+};
+
+export const chrome = {
+  identity: identity,
+};
