+++ jsconfig.json
@@ -1,0 +1,5 @@
+{
+  "typeAcquisition": {
+    "include": ["jest"]
+  }
+}
