+++ package.json
@@ -27,7 +27,6 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "testw": "jest --watchAll",
     "test": "jest",
     "test1": "react-scripts test",
     "eject": "react-scripts eject",
@@ -60,7 +59,7 @@
   },
   "jest": {
     "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/chromeMock.js",
       "\\.(css|less|scss|sass)$": "identity-obj-proxy"
     }
   }
