+++ public/animalGenerator.js
@@ -1,4 +1,4 @@
-let animals = [
+export let animals = [
   "airbaloon",
   "android",
   "arcade",
@@ -54,7 +54,7 @@ let animals = [
 /**
  * Distributes animal for new user
  */
-function getAnimal(animalsLeft) {
+export function getAnimal(animalsLeft) {
   if (animalsLeft.length === 0) {
     throw "ERROR";
   }
@@ -67,7 +67,7 @@ function getAnimal(animalsLeft) {
  * Puts animal back if a user leaves team
  * @param {string} animal
  */
-function addAnimal(animalsLeft, animal) {
+export function addAnimal(animalsLeft, animal) {
   if (animals.includes(animal)) {
     animalsLeft.push(animal);
   }
