+++ public/background.html
@@ -8,18 +8,15 @@
   <script src="https://www.gstatic.com/firebasejs/7.14.2/firebase-app.js"></script>
   <script src="https://www.gstatic.com/firebasejs/7.14.2/firebase-firestore.js"></script>
 
-  <script src="firebaseInit.js"></script>
+  <script type="module" src="firebaseInit.js"></script>
   <script type="text/javascript" src="constants.js"></script>
 
-  <!-- <script src="realTime.js"></script> -->
-  <script src="animalGenerator.js"></script>
-  <script src="github-oauth.js"></script>
-  <script src="githubTracker.js"></script>
-  <script src="background.js"></script>
-  <!-- <script src="tabTime.js"></script> -->
-  <!-- <script src="realTime.js"></script> -->
-
-  <!-- <script src="/oauth/oauth2.js"></script> -->
+  <script type="module" src="animalGenerator.js"></script>
+  <script type="module" src="github-oauth.js"></script>
+  <script type="module" src="githubTracker.js"></script>
+  <script type="module" src="userAndTeams.js"></script>
+  <script type="module" src="tabs.js"></script>
+  <script type="module" src="background.js"></script>
 
   <body></body>
 </html>
