+++ public/github-oauth.js
@@ -1,41 +1,42 @@
-'use strict';
+/* global chrome */
 
-var signin_button;
-var revoke_button;
-var user_info_div;
+let signin_button;
+let revoke_button;
+let user_info_div;
 
-var tokenFetcher = (function() {
+let tokenFetcher = (function () {
   // Replace clientId and clientSecret with values obtained by you for your
   // application https://github.com/settings/applications.
-  var clientId = '825ffe9603a27981677a';
+  let clientId = "825ffe9603a27981677a";
   // Note that in a real-production app, you may not want to store
   // clientSecret in your App code.
-  var clientSecret = '0c4548e8c7cb92a85f868a14431d4f1fd29106ea';
-  var redirectUri = chrome.identity.getRedirectURL('provider_cb');
-  var redirectRe = new RegExp(redirectUri + '[#\?](.*)');
+  let clientSecret = "0c4548e8c7cb92a85f868a14431d4f1fd29106ea";
+  let redirectUri = chrome.identity.getRedirectURL("provider_cb");
+  let redirectRe = new RegExp(redirectUri + "[#?](.*)");
 
-  var access_token = null;
-  localStorage.removeItem("token")
+  let access_token = null;
+  localStorage.removeItem("token");
 
   return {
-    getToken: function(interactive, callback) {
+    getToken: function (interactive, callback) {
       // In case we already have an access_token cached, simply return it.
       if (access_token) {
         callback(null, access_token);
         return;
       }
 
-      var options = {
-        'interactive': interactive,
-        'url': 'https://github.com/login/oauth/authorize' +
-                '?client_id=' + clientId + "&scope=repo"+
-                '&redirect_uri=' + encodeURIComponent(redirectUri)
-      }
+      let options = {
+        interactive: interactive,
+        url:
+          "https://github.com/login/oauth/authorize" +
+          "?client_id=" +
+          clientId +
+          "&scope=repo" +
+          "&redirect_uri=" +
+          encodeURIComponent(redirectUri),
+      };
 
-      chrome.identity.launchWebAuthFlow(options, function(redirectUri) {
-        // console.log('launchWebAuthFlow completed', chrome.runtime.lastError,
-            // redirectUri);
-
+      chrome.identity.launchWebAuthFlow(options, function (redirectUri) {
         if (chrome.runtime.lastError) {
           callback(new Error(chrome.runtime.lastError));
           return;
@@ -46,19 +47,18 @@ var tokenFetcher = (function() {
         //     &refresh_token={value}
         // or:
         // https://{app_id}.chromiumapp.org/provider_cb#code={value}
-        var matches = redirectUri.match(redirectRe);
+        let matches = redirectUri.match(redirectRe);
         if (matches && matches.length > 1)
           handleProviderResponse(parseRedirectFragment(matches[1]));
-        else
-          callback(new Error('Invalid redirect URI'));
+        else callback(new Error("Invalid redirect URI"));
       });
 
       function parseRedirectFragment(fragment) {
-        var pairs = fragment.split(/&/);
-        var values = {};
+        let pairs = fragment.split(/&/);
+        let values = {};
 
-        pairs.forEach(function(pair) {
-          var nameval = pair.split(/=/);
+        pairs.forEach(function (pair) {
+          let nameval = pair.split(/=/);
           values[nameval[0]] = nameval[1];
         });
 
@@ -66,109 +66,111 @@ var tokenFetcher = (function() {
       }
 
       function handleProviderResponse(values) {
-        // console.log('providerResponse', values);
-        if (values.hasOwnProperty('access_token'))
+        if (values.hasOwnProperty("access_token"))
           setAccessToken(values.access_token);
         // If response does not have an access_token, it might have the code,
         // which can be used in exchange for token.
-        else if (values.hasOwnProperty('code'))
+        else if (values.hasOwnProperty("code"))
           exchangeCodeForToken(values.code);
-        else 
-          callback(new Error('Neither access_token nor code avialable.'));
+        else callback(new Error("Neither access_token nor code avialable."));
       }
 
       function exchangeCodeForToken(code) {
-        var xhr = new XMLHttpRequest();
-        xhr.open('GET',
-                  'https://github.com/login/oauth/access_token?' +
-                  'client_id=' + clientId +
-                  '&client_secret=' + clientSecret +
-                  '&redirect_uri=' + redirectUri +
-                  '&code=' + code);
-        xhr.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded');
-        xhr.setRequestHeader('Accept', 'application/json');
+        let xhr = new XMLHttpRequest();
+        xhr.open(
+          "GET",
+          "https://github.com/login/oauth/access_token?" +
+            "client_id=" +
+            clientId +
+            "&client_secret=" +
+            clientSecret +
+            "&redirect_uri=" +
+            redirectUri +
+            "&code=" +
+            code
+        );
+        xhr.setRequestHeader(
+          "Content-Type",
+          "application/x-www-form-urlencoded"
+        );
+        xhr.setRequestHeader("Accept", "application/json");
         xhr.onload = function () {
           // When exchanging code for token, the response comes as json, which
           // can be easily parsed to an object.
           if (this.status === 200) {
-            var response = JSON.parse(this.responseText);
-            // console.log(response);
+            let response = JSON.parse(this.responseText);
 
-            if (response.hasOwnProperty('access_token')) {
-                setAccessToken(response.access_token);
-                // let item = { "token" : response.access_token };
-                localStorage.setItem("token", response.access_token);
-                
-                // signin_button.disabled = true;
+            if (response.hasOwnProperty("access_token")) {
+              setAccessToken(response.access_token);
+              // let item = { "token" : response.access_token };
+              localStorage.setItem("token", response.access_token);
+
+              // signin_button.disabled = true;
             } else {
-              callback(new Error('Cannot obtain access_token from code.'));
+              callback(new Error("Cannot obtain access_token from code."));
             }
           } else {
-            console.log('code exchange status:', this.status);
-            callback(new Error('Code exchange failed'));
+            console.log("code exchange status:", this.status);
+            callback(new Error("Code exchange failed"));
           }
         };
         xhr.send();
       }
 
       function setAccessToken(token) {
-        access_token = token; 
-        // console.log('Setting access_token: ', access_token);
+        access_token = token;
         callback(null, access_token);
       }
     },
-    removeCachedToken: function(token_to_remove) {
-      if (access_token == token_to_remove)
-        access_token = null;
-    }
-  }
+    removeCachedToken: function (token_to_remove) {
+      if (access_token == token_to_remove) access_token = null;
+    },
+  };
 })();
 
-
 function xhrWithAuth(method, url, interactive, callback) {
-    var retry = true;
-    var access_token;
+  let retry = true;
+  let access_token;
 
-    // console.log('xhrWithAuth', method, url, interactive);
-    getToken();
+  getToken();
 
-    function getToken() {
-        tokenFetcher.getToken(interactive, function(error, token) {
-            // console.log('token fetch', error, token);
-            if (error) {
-                callback(error);
-                return;
-            }
+  function getToken() {
+    tokenFetcher.getToken(interactive, function (error, token) {
+      if (error) {
+        callback(error);
+        return;
+      }
 
-            access_token = token;
-            // console.log(access_token)
-        });
-    }
+      access_token = token;
+    });
+  }
 }
 
-function getUserInfo(interactive) {
-  xhrWithAuth('GET',
-              'https://api.github.com/user',
-              interactive,
-              onUserInfoFetched);
+export function getUserInfo(interactive) {
+  xhrWithAuth(
+    "GET",
+    "https://api.github.com/user",
+    interactive,
+    onUserInfoFetched
+  );
 }
 
 // Functions updating the User Interface:
 function showButton(button) {
-  if(button != null){
-    button.style.display = 'inline';
+  if (button != null) {
+    button.style.display = "inline";
     button.disabled = false;
   }
 }
 
 function hideButton(button) {
-  if(button != null){
-    button.style.display = 'none';
+  if (button != null) {
+    button.style.display = "none";
   }
 }
 
 function disableButton(button) {
-  if(button != null){
+  if (button != null) {
     button.disabled = true;
   }
 }
@@ -176,101 +178,94 @@ function disableButton(button) {
 function onUserInfoFetched(error, status, response) {
   if (!error && status == 200) {
     console.log("Got the following user info: " + response);
-    var user_info = JSON.parse(response);
+    let user_info = JSON.parse(response);
     populateUserInfo(user_info);
     showButton(revoke_button);
     fetchUserRepos(user_info["repos_url"]);
   } else {
-    console.log('infoFetch failed', error, status);
+    console.log("infoFetch failed", error, status);
     showButton(signin_button);
   }
 }
 
 function populateUserInfo(user_info) {
-  var elem = user_info_div;
-  var nameElem = document.createElement('div');
-  nameElem.innerHTML = "<b>Hello " + user_info.name + "</b><br>"
-      + "Your github page is: " + user_info.html_url;
+  let elem = user_info_div;
+  let nameElem = document.createElement("div");
+  nameElem.innerHTML =
+    "<b>Hello " +
+    user_info.name +
+    "</b><br>" +
+    "Your github page is: " +
+    user_info.html_url;
   elem.appendChild(nameElem);
 }
 
 function fetchUserRepos(repoUrl) {
-  xhrWithAuth('GET', repoUrl, false, onUserReposFetched);
+  xhrWithAuth("GET", repoUrl, false, onUserReposFetched);
 }
 
 function onUserReposFetched(error, status, response) {
-  var elem = document.querySelector('#user_repos');
-  elem.value='';
+  let elem = document.querySelector("#user_repos");
+  elem.value = "";
   if (!error && status == 200) {
     console.log("Got the following user repos:", response);
-    var user_repos = JSON.parse(response);
-    user_repos.forEach(function(repo) {
+    let user_repos = JSON.parse(response);
+    user_repos.forEach(function (repo) {
       if (repo.private) {
         elem.value += "[private repo]";
       } else {
         elem.value += repo.name;
       }
-      elem.value += '\n';
+      elem.value += "\n";
     });
   } else {
-    console.log('infoFetch failed', error, status);
+    console.log("infoFetch failed", error, status);
   }
-  
 }
 
 // Handlers for the buttons's onclick events.
 function interactiveSignIn() {
-    tokenFetcher.getToken(true, function(error, access_token) {
-      if (error) {
-          document.querySelector('#signin').innerHTML='SIGN IN';
-          showButton(signin_button);
-      } else {
-          document.querySelector('#signin').innerHTML='&#10004';
-          disableButton(signin_button);
-          // console.log(access_token);
+  tokenFetcher.getToken(true, function (error, access_token) {
+    if (error) {
+      document.querySelector("#signin").innerHTML = "SIGN IN";
+      showButton(signin_button);
+    } else {
+      document.querySelector("#signin").innerHTML = "&#10004";
+      disableButton(signin_button);
       //   getUserInfo(true);
-      }
-    });
+    }
+  });
 }
 
 function revokeToken() {
   // We are opening the web page that allows user to revoke their token.
-  window.open('https://github.com/settings/applications');
+  window.open("https://github.com/settings/applications");
   // And then clear the user interface, showing the Sign in button only.
   // If the user revokes the app authorization, they will be prompted to log
   // in again. If the user dismissed the page they were presented with,
   // Sign in button will simply sign them in.
   // user_info_div.textContent = '';
   hideButton(revoke_button);
-  document.querySelector('#signin').innerHTML='SIGN IN';
+  document.querySelector("#signin").innerHTML = "SIGN IN";
   showButton(signin_button);
-  if(signin_button != null){
+  if (signin_button != null) {
     signin_button.onclick = interactiveSignIn;
   }
 }
 
-    
-  signin_button = document.querySelector('#signin');
-  if(signin_button != null){
-    signin_button.onclick = interactiveSignIn;
-  }
-  revoke_button = document.querySelector('#revoke');
-  if(revoke_button != null){
-    revoke_button.onclick = revokeToken;
-  }
+signin_button = document.querySelector("#signin");
+if (signin_button != null) {
+  signin_button.onclick = interactiveSignIn;
+}
+revoke_button = document.querySelector("#revoke");
+if (revoke_button != null) {
+  revoke_button.onclick = revokeToken;
+}
 
-  // user_info_div = document.querySelector('#user_info');
+// user_info_div = document.querySelector('#user_info');
 
-  // console.log(signin_button, revoke_button, user_info_div);
+if (signin_button != null) {
+  showButton(signin_button);
+}
 
-  if(signin_button != null) {
-    showButton(signin_button);
-  }
-  
-  getUserInfo(false);
-
-
-
-
-  
-  
+getUserInfo(false);
