+++ public/githubTracker.js
@@ -1,3 +1,7 @@
+/* global firebase */
+import { db } from "./firebaseInit.js";
+import { getUserInfo } from "./github-oauth.js";
+import { userEmail } from "./userAndTeams.js";
 /**
  *  Fetches data from github using the Github REST API. Header
  * includes the token which was saved after user sign-in with Github.
@@ -133,8 +137,7 @@ async function getMostRecentCommit() {
  * @param {string} teamCode
  * @param {string} animal
  */
-function sendToDB(teamCode, animal) {
-  // console.log(teamCode)
+export function sendToDB(teamCode, animal) {
   if (teamCode === undefined || animal === undefined) {
     return;
   }
@@ -146,14 +149,11 @@ function sendToDB(teamCode, animal) {
     getMostRecentCommit()
       .then((arr) => {
         let max = "00:00:00";
-        // console.log(arr);
         arr.forEach((e) => {
           if (e > max) {
             max = e;
           }
         });
-
-        // console.log(max);
 
         if (max != "00:00:00") {
           let msg = {
@@ -169,10 +169,8 @@ function sendToDB(teamCode, animal) {
             localStorage.setItem("oldElements", JSON.stringify(item));
 
             // chrome.runtime.sendMessage(msg, function (response) {
-            //     console.log(response);
             //     resolve(response);
             // });
-            console.log("here");
             // let seconds = JSON.parse(currData)[tabUrl] / 1000;
             let seconds = new Date().toLocaleTimeString();
             //let seconds =
@@ -211,7 +209,6 @@ function sendToDB(teamCode, animal) {
               );
           } else {
             let item = { url: "github.com", time: max };
-            // console.log("Local Storage is not empty...")
             let oldElements = JSON.parse(localStorage.getItem("oldElements"));
 
             // Checks if an item exists in the local storage
@@ -224,16 +221,10 @@ function sendToDB(teamCode, animal) {
             console.log(itemExists);
 
             if (itemExists) {
-              // console.log("Item is in Local Storage...")
             } else {
               console.log("Item is not in Local Storage...");
               oldElements.push(item);
               localStorage.setItem("oldElements", JSON.stringify(oldElements));
-              // chrome.runtime.sendMessage(msg, function (response) {
-              //     console.log("send message");
-              //     console.log(response);
-              //     resolve(response);
-              // });
               console.log("here");
               // let seconds = JSON.parse(currData)[tabUrl] / 1000;
               let seconds = new Date().toLocaleTimeString();
