+++ src/App.js
@@ -10,7 +10,6 @@ import Teams from "./components/Teams";
 import Home from "./components/Home";
 import Timeline from "./components/Timeline";
 import Charts from "./components/Charts";
-import TimelineDemo from "./components/TimelineDemo";
 
 // for Testing purpose
 class App extends Component {
@@ -37,11 +36,6 @@ class App extends Component {
                 <Route exact path="/teams" component={Teams}></Route>
                 <Route exact path="/timeline" component={Timeline}></Route>
                 <Route exact path="/charts" component={Charts}></Route>
-                <Route
-                  exact
-                  path="/timelinedemo"
-                  component={TimelineDemo}
-                ></Route>
                 <Route path="/" component={Home}></Route>
               </Switch>
             </div>
