+++ src/components/Timeline.js
@@ -64,7 +64,6 @@ export default class Timeline extends Component {
         timeline: request.message.timeWasted,
       });
 
-    console.log(request);
     return true;
   };
   /**
